# Removal of Bootstrap Script Execution State and User Config Propagation State
Currently, Minitrino writes a checksum representing the contents of a bootstrap
script that was executed inside of a container to
`/opt/minitrino/bootstrap-status.txt`. Additionally, user-provided configs
(supplied either through environment variables or through a Docker Compose
environment variable) are recorded in `/opt/minitrino/user-config.txt`. When
containers are redeployed through a subsequent `minitrino provision` command,
Minitrino checks if a given bootstrap's checksum or a set of user config's
checksum matches a checksum previously recorded in one of the two files.

The only scenarios this would ever be useful for since adding `--force-recreate`
to the Compose command are:

1. A bootstrap script modifies data stored in a *named volume* that does not
   check whether its logic has already been executed.
2. A container backed by a *named volume* carries out procedures that break when
   the container is restarted, similarly because it does not check whether its
   startup process has already been completed.

The first scenario can be easily avoided by simply checking if the bootstrap
script has already executed before executing it again. Again, this is only
necessary if the script is affecting data stored in a named volume, as named
volumes aren't automatically deleted when the associated container is recreated.

The second scenario is unlikely to occur, and if it does, either (1) don't use
the image or (2) don't back the container with a named volume.

The `--force-recreate` option causes every container in the Compose app to
recreate whenever a subsequent `docker compose up` command is executed. When
containers are recreated, they will boot back up and (1) execute their default
command (e.g. `exec /bin/start-hms.sh`) and (2) execute any bootstrap scripts
passed in by Minitrino. Even though this takes a little more time, it is
worthwhile to ensure that all containers are up-to-date. Prior to [parallelizing
container restarts after bootstrap
execution](https://github.com/jefflester/minitrino/commit/6d12e46ed0c67469503ef93ba5d92a7306fb2f38),
this took way too long. Now, time is not as much of an issue.